moved user tests from users_controller_spec to users_spec for integration testing
