+++ dispatchr-ruby/app/controllers/application_controller.rb
@@ -3,32 +3,6 @@ require "jwt"
 class ApplicationController < ActionController::API
   attr_reader :current_user
 
-  def create
-    if User.find_by(email: params[:email]).present?
-      render status: :bad_request, json: {error: "User already exists"} and return
-    end
-    
-    address = Address.new(
-        address: params[:address][:address],
-        latitude: params[:address][:latitude],
-        longitude: params[:address][:longitude]
-    )
-
-    @user = User.create(
-        name: params[:name],
-        username: params[:username],
-        email: params[:email],
-        password: params[:password],
-        address: address
-    )
-    if @user
-      token = Authentication::JsonWebToken.encode(@user.id)
-      render json: { user: @user, auth_token: token }
-    else
-      render json: @user.errors, status: :bad_request
-    end
-  end
-
   protected
   def authenticate_request!
     unless user_id_in_token?
